+++ lib/steel.h
@@ -2,7 +2,7 @@
 #include <array>
 #include <cstddef>
 
-using data_type = float;
+using data_type = double;
 
 namespace steel {
